Relocate specs -> tests
